Rewrite account creation wizard UI code (emailWizard.js). Bug 549045, r=bwinton, ui-r=clarkbw, a=Standard8
The previous UI code had a bit the taste of spaghetti bolognese. Delicious, but hard to follow.

Concrete problem was mostly when you did not want autodetection, but manual config:
1. The Stop button didn't work, it continued for a bit after you pressed Stop
2. When you pressed "Manual Config", we immediately created the account and went into the
   account manager, instead of going into the inline "Edit" mode.
3. Worse, the account was created with whatever server happened to be tested last,
   so when pressing "Manual config", you got a litterally random config.
   You couldn't change IMAP<->POP3 later, either.
4. The inline Edit didn't take the values you gave it as given, but second-guessed them
   upon testing, and felt free to change them, even if the user explicitly said otherwise.

Understandably, that infuriated some users.
This change should fix all that.
The stop button now actually works, and the "Manual config" button goes into "Edit" mode.

This also uses a different philosophy, in a way:
After detection, we have a "result" screen, where the user is only *presented*
the values that we found, but he cannot *change* them (apart from IMAP vs. POP3).
The result screen is only to confirm that the servers are good. It should be
very quick and easy to read.
If the user wants to change the values, we make that easy with the "manual config"
button (formerly "Edit") and going into a different where we present all the values
as input fields. There, we take what the user said as a given command.
But the default is "Auto", where we can guess, so you have a combination of both
user being in command and us supporting the user.
Only if the user needs to change some advanced values like IMAP root folder, and
wants to do so immediately, we have the "Advanced config" button, which does the
same as the old "manual config".

This was a lot of work to code, and then even more with a one-year foray to get
this reviewed, so I really hope this helps.

Permission to commit into CLOSED TREE by Standard8/9 on IRC.

X-Channel-Repo: comm-central
X-Channel-Converted-Revision: 75875cb17fb0a20caabb627fe743b211f8d43eaa