Force access keys to Cyrillic in Beta.
Used modified silme script from Gandalf.  The script forced the access keys
into cyrillic, but also re-sorted keys in the output file, amounting to a
change that is difficult, if at all possible, to patch into Aurora.

Though Axel suggested a backport, I think this is already quirky enough so
will not backport.  Instead, i'll do the same on Aurora.