Bug 994055 - Part 1: Add a toggle sidebar panel button to the inspector. r=miker
This adds a new icon, next to the searchbox (like in the debugger) to expand
or collapse the sidebar panel in the inspector.
The state is *not* persisted yet when the toolbox is closed.
The button is hidden when the toolbox host switches to 'side'.

X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: d34c868e5f6707281fb2f44cc3576dcadda3740a