Bug 1301305 - Add a performance warning type for transform animations that should be synchronized with geometric animations; r=hiro, r=flod
This patch adds a new performance warning type for the case when we start
a transform animation at the same time as an animation that includes a
geometric property. In that case we run the transform animation on the main
thread so that it is synchronized with the geometric animation (which we can
only run on the main thread).

This differs from CompositorAnimationWarningTransformWithGeometricProperties
in that this applies across different elements whilst the existing warning
only covers the case when the same animation animates both transform and
geometric properties.

MozReview-Commit-ID: EcOMo4VDAYY
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: d14b2103446d083f72f5262d79168e8eb539f401