Bug 1322873 - Bug 1322873 - Add React event bubbles to the Markup View r=tromey
DevRel have made it clear that one of the number one complaints they hear is that we are not supporting React in our tools.

So how about we have our event bubbles include React events and allow people to go to the event listener source in the debugger?

I don't believe that any other tool does this so it is totally worth doing... and in time for Christmas as well ;)

Works just fine in development and production versions of React.

It also works in the browser toolbox so it can be used to debug events in our own tools e.g. The Debugger.

The files under devtools/client/inspector/markup/test/ are either test or React library files so they only really need a cursory glance.

This means that you should focus on the following files when reviewing:
 - devtools/client/locales/en-US/inspector.properties
 - devtools/client/shared/widgets/tooltip/EventTooltipHelper.js
 - devtools/server/actors/inspector.js
 - devtools/server/event-parsers.js

Now allowed the use of JSX in mochitests and fixed all eslint errors.

MozReview-Commit-ID: AtxhainieQe
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: ff9a07a874fd1ef273f23aa26cac61a95ba0f342