Bug 1315201: Add activity_stream_welcome_dismiss str. r=liuche
activity_stream_dismiss was being used to dismiss the welcome screen before but
this string is intended to be used in a different location for a different
purpose and thus should never have been set to the same string.

It's necessary to change the String here so we can modify the contents of the
activity_stream_dismiss String.

MozReview-Commit-ID: 87zaHQ5TJjN
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: 9c6a40a07c690e86a1b9998b3794acec5a86dc7a