Bug 1388377: Rm WelcomePanelRow. r=liuche
We show a message:
1) On first run
2) When there are no highlights

However, these are actually the same case (there are no highlights) so I wanted
to defer to that situation.

I felt it was easier to rm the WelcomePanel and rewrite it than it was to try
to morph it into an empty state for the highlights.

MozReview-Commit-ID: CrRbzA0NoRx
X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: c3b570626261a5bd5a8b9aca805935b72e320c55