Bug 1551586 - Display a warning when min-width, min-height etc. are used incorrectly. r=miker,fluent-reviewers,flod.
This introduce a new form of invalid messages. Until now, all the
messages were like "X has no effect on this since it's not Y".
But in this case, the width/height properties applies to all but
a few cases, which means we can't really keep the same shape of
message (or it would be "since it's not A, B, C, D, ...).
So we're switching to a message that prints the display property
of the element ("X has no effect on this element since it has a
display of Y").
In order to do that, we need to pipe the element computed display
into the inactive tooltip.

Differential Revision: https://phabricator.services.mozilla.com/D32805

X-Channel-Repo: mozilla-central
X-Channel-Converted-Revision: 31cec3975dd0988d798cb63711477e9745339064
X-Channel-Repo: releases/mozilla-beta
X-Channel-Revision: 4e5980a74c9c5023f5f64741a4bee1c1055af24c
X-Channel-Repo: releases/mozilla-release
X-Channel-Revision: df853c362664989e025d1206b59c37c435186a91